+++ network/views/wifi.py
@@ -1,8 +1,7 @@
 from django.shortcuts import get_object_or_404, redirect, render
 
-from ..forms import WifiForm
+from ..forms.networking import WifiForm
 from ..models import WifiNetwork
-
 
 def index(request):
     wifis = WifiNetwork.objects.all()
